展示menuCreate测试页 menuCreate.jsp 接收create创建菜单的参数转换为String,内部进行JSONObject的转换 对delete进行了判定是否为空的操作(0,无菜单可删除; 1删除菜单成功) 加入servlet路径以及jsp解析前后缀 tomcat-embed-core-9.0.13依然报错. 前端页面menuCreate.jsp

+++ src/main/java/com/baigehuidi/demo/controller/WeixinMenuController.java
@@ -1,10 +1,8 @@
 package com.baigehuidi.demo.controller;
 
 import com.alibaba.fastjson.JSONObject;
-import com.baigehuidi.demo.loader.MenuComponentInsLoader;
-import com.baigehuidi.demo.loader.WeixinInsLoader;
 import com.baigehuidi.demo.service.WeixinComponentService;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.weixin4j.WeixinException;
@@ -15,17 +13,26 @@ import org.weixin4j.model.menu.Menu;
 @RequestMapping("/menu")
 public class WeixinMenuController {
 
+    //获取menuComponent实例,源码中没有@Component
     private static MenuComponent menuComponent = WeixinComponentService.getMenuComponent();
 
     /**
      * 需要前端页面form表单进行自定义菜单数据传输 TODO
-     * @param jsonObject
+     * @param  menuJson
      */
-    @RequestMapping("create")
-    public void createMenu(JSONObject jsonObject) {
 
+
+    @RequestMapping("create")
+    public void createMenu(@RequestBody String menuJson) {
+
+//        System.out.println(menuJson);
+        //change menuJson -> (JSONObject)menuJson
+        JSONObject jsonObject = (JSONObject) JSONObject.parse(menuJson);
+//        System.out.println(jsonObject);
+        //new a Menu instance
         Menu menu = new Menu(jsonObject);
         try {
+            //create the menu
             menuComponent.create(menu);
         } catch (WeixinException e) {
             e.printStackTrace();
@@ -39,6 +46,7 @@ public class WeixinMenuController {
         try {
             return menuComponent.get();
         } catch (WeixinException e) {
+            System.out.println("获取失败,菜单可能不存在.");
             e.printStackTrace();
         } finally {
         }
@@ -46,14 +54,21 @@ public class WeixinMenuController {
     }
 
     @RequestMapping("delete")
-    public void deleteMenu() {
+    public Integer deleteMenu() {
         try {
+            if(getMenu()==null){
+                //如果菜单本来就是null,那么返回0
+                System.out.println("自定义菜单为null~");
+                return 0;
+            }
             menuComponent.delete();
         } catch (WeixinException e) {
             e.printStackTrace();
         } finally {
         }
         System.out.println("deleteMenu ok.");
+        //有菜单被删除,返回1
+        return 1;
     }
 
 }
